# api/openapi.yaml
openapi: 3.0.3
info:
  title: Waterflow API
  version: 1.1.0
  description: |
    声明式工作流编排引擎 REST API
    
    Waterflow 提供 YAML 驱动的分布式工作流编排能力，支持跨服务器任务执行、
    自动重试、完整的状态持久化和实时日志流。
    
    ## 架构更新 (ADR-0009)
    
    根据 [ADR-0009: 工作流定义与执行分离](../docs/adr/0009-workflow-definition-execution-separation.md)，
    API 已进行语义重构：
    
    - **Definition API** (`/v1/workflows`): 管理工作流定义（YAML + 元数据）
    - **Execution API** (`/v1/executions`): 管理执行实例
    - **Schedule API** (`/v1/workflows/{name}/schedules`): 挂载到定义下
    - **Webhook API** (`/v1/workflows/{name}/webhooks`): 挂载到定义下
    
    完整 API 清单详见 [api-inventory.md](../docs/api-inventory.md)。
    
    > **注意:** 当前 OpenAPI 规范尚未完全同步 ADR-0009 变更，将在后续版本更新。
    > 参考文档以 api-inventory.md 为准。
    
    ## 核心功能
    - 🎯 声明式 YAML DSL (类似 GitHub Actions)
    - 🔄 Event Sourcing 架构 - 完整执行历史追踪
    - 🌐 分布式 Agent 系统 - 跨服务器任务路由
    - 🔌 插件化节点系统 - 10+ 内置节点，支持自定义
    - 📊 实时日志流 - 支持 SSE 和 NDJSON 格式
    
    ## 认证
    当前版本支持两种认证方式:
    - **Bearer Token**: JWT 格式，用于用户认证
    - **API Key**: 在 HTTP Header 中传递 `X-API-Key`
    
    开发环境默认不需要认证。生产环境请配置 TLS 和认证。
  contact:
    name: Websoft9 Team
    url: https://github.com/websoft9/waterflow
    email: support@websoft9.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.waterflow.example.com
    description: 生产环境
...
security:
  - bearerAuth: []
  - apiKey: []
  - {} # Allow unauthenticated access for development
paths:
...
  /v1/workflows:
    post:
      summary: 提交工作流
      description: |
        提交 YAML 格式的工作流定义并开始执行。
        
        工作流将被解析、验证后提交到 Temporal 引擎执行。
        返回工作流 ID 用于后续状态查询和日志获取。
      tags: [Workflows]
      operationId: submitWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWorkflowRequest'
            examples:
              hello-world:
                summary: Hello World 示例
                value:
                  yaml: |
                    name: hello-world
                    jobs:
                      greet:
                        runs-on: default
                        steps:
                          - name: Say Hello
                            uses: shell@v1
                            with:
                              command: echo
                              args: ["Hello, World!"]
              with-variables:
                summary: 使用变量示例
                value:
                  yaml: |
                    name: deploy-app
                    jobs:
                      deploy:
                        runs-on: web-servers
                        steps:
                          - uses: shell@v1
                            with:
                              command: docker
                              args: ["pull", "${{ vars.image }}"]
                  vars:
                    image: "nginx:latest"
      responses:
        '201':
          description: 工作流创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitWorkflowResponse'
              example:
                id: "wf_20260115_abc123"
                run_id: "def456ghi789"
                name: "hello-world"
                status: "pending"
                created_at: "2026-01-15T10:30:00Z"
                url: "http://localhost:8088/namespaces/default/workflows/wf_20260115_abc123/def456ghi789"
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    
    get:
      summary: 列出工作流
      description: |
        查询工作流列表，支持分页和过滤。
        
        可按状态、名称、创建时间等条件过滤。
      tags: [Workflows]
      operationId: listWorkflows
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          description: 按状态过滤
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: name
          in: query
          description: 按工作流名称模糊搜索
          schema:
            type: string
        - name: created_after
          in: query
          description: 创建时间起始（ISO 8601 格式）
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: 创建时间结束（ISO 8601 格式）
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 成功返回工作流列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
              example:
                workflows:
                  - id: "wf_20260115_abc123"
                    name: "hello-world"
                    status: "completed"
                    created_at: "2026-01-15T10:30:00Z"
                    completed_at: "2026-01-15T10:30:05Z"
                  - id: "wf_20260115_xyz789"
                    name: "deploy-app"
                    status: "running"
                    created_at: "2026-01-15T10:32:00Z"
                meta:
                  total: 42
                  page: 1
                  limit: 20
                  total_pages: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{id}:
    get:
      summary: 查询工作流状态
      description: |
        获取指定工作流的详细状态信息。
        
        包含所有 Job 和 Step 的执行状态、输出、错误信息等。
      tags: [Workflows]
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200':
          description: 成功返回工作流详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetailResponse'
              example:
                id: "wf_20260115_abc123"
                run_id: "def456ghi789"
                name: "hello-world"
                status: "completed"
                created_at: "2026-01-15T10:30:00Z"
                started_at: "2026-01-15T10:30:01Z"
                completed_at: "2026-01-15T10:30:05Z"
                jobs:
                  - name: "greet"
                    status: "completed"
                    started_at: "2026-01-15T10:30:01Z"
                    completed_at: "2026-01-15T10:30:05Z"
                    steps:
                      - name: "Say Hello"
                        status: "completed"
                        started_at: "2026-01-15T10:30:02Z"
                        completed_at: "2026-01-15T10:30:04Z"
                        outputs:
                          stdout: "Hello, World!\n"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{id}/logs:
    get:
      summary: 获取工作流日志
      description: |
        获取工作流执行日志，支持 NDJSON 流式输出。
        
        可按日志级别、Job、Step 过滤，支持 tail 模式实时跟踪。
      tags: [Workflows]
      operationId: getWorkflowLogs
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - name: level
          in: query
          description: 日志级别过滤
          schema:
            type: string
            enum: [debug, info, warn, error]
        - name: job
          in: query
          description: 按 Job 名称过滤
          schema:
            type: string
        - name: step
          in: query
          description: 按 Step 名称过滤
          schema:
            type: string
        - name: tail
          in: query
          description: 只返回最后 N 条日志
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1000
      responses:
        '200':
          description: 成功返回日志（NDJSON 格式）
          content:
            application/x-ndjson:
              schema:
                type: string
              example: |
                {"timestamp":"2026-01-15T10:30:02Z","level":"info","job":"greet","step":"Say Hello","message":"Executing command: echo"}
                {"timestamp":"2026-01-15T10:30:04Z","level":"info","job":"greet","step":"Say Hello","message":"Command output: Hello, World!"}
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{id}/cancel:
    post:
      summary: 取消工作流
      description: |
        取消正在运行的工作流。
        
        已完成或失败的工作流无法取消。
      tags: [Workflows]
      operationId: cancelWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200':
          description: 取消请求已接受
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workflow cancellation requested"
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{id}/terminate:
    post:
      summary: 强制终止工作流
      description: |
        立即强制终止工作流，不执行清理逻辑。
        
        **Terminate vs Cancel 区别:**
        - **Terminate**: 立即强制终止，不触发 defer 或 cleanup activities
        - **Cancel**: 优雅取消，允许 cleanup logic 执行
        
        适用场景: 工作流卡住或需要紧急回滚。
      tags: [Workflows]
      operationId: terminateWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 终止原因 (可选)
                  example: "Deployment rollback required"
      responses:
        '204':
          description: 工作流已终止 (无响应体)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{id}/rerun:
    post:
      summary: 重新运行工作流
      description: |
        基于现有工作流创建新的执行实例。
        
        保留原工作流的 YAML 定义和变量。
      tags: [Workflows]
      operationId: rerunWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '201':
          description: 新工作流创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitWorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ==================== Validation ====================
  /v1/validate:
    post:
      summary: 验证工作流 YAML
      description: |
        在提交前验证工作流 YAML 语法和语义。
        
        检查语法错误、节点参数、表达式语法等。
      tags: [Validation]
      operationId: validateWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [yaml]
              properties:
                yaml:
                  type: string
                  description: 工作流 YAML 内容
            example:
              yaml: |
                name: test-workflow
                jobs:
                  test:
                    runs-on: default
                    steps:
                      - uses: shell@v1
                        with:
                          command: echo test
      responses:
        '200':
          description: 验证通过
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Validation successful"
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ==================== Schedules ====================
  /v1/workflows/{name}/schedules:
    post:
      summary: 创建工作流调度
      description: |
        为指定工作流创建定时调度。
        
        支持 Cron 表达式、时区设置和自定义变量绑定。
        调度创建后将由 Temporal 管理生命周期。
      tags: [Schedules]
      operationId: createSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: 调度创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    get:
      summary: 列出工作流调度
      description: |
        获取指定工作流的所有调度列表。
        
        支持分页和状态过滤。
      tags: [Schedules]
      operationId: listSchedules
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          description: 按状态过滤
          schema:
            type: string
            enum: [active, paused]
      responses:
        '200':
          description: 成功返回调度列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedules:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduleResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{name}/schedules/{schedule_id}:
    get:
      summary: 获取调度详情
      description: |
        获取指定调度的详细信息。
        
        包含下次执行时间、历史记录等。
      tags: [Schedules]
      operationId: getSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      responses:
        '200':
          description: 成功返回调度详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      summary: 更新调度配置
      description: |
        更新调度的 Cron 表达式、时区或变量。
        
        更新后立即生效。
      tags: [Schedules]
      operationId: updateSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      summary: 删除调度
      description: |
        删除指定调度。
        
        调度删除后，已触发的工作流不受影响。
      tags: [Schedules]
      operationId: deleteSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{name}/schedules/{schedule_id}/pause:
    post:
      summary: 暂停调度
      description: |
        暂停调度，不再触发新的工作流执行。
        
        可选提供暂停原因。
      tags: [Schedules]
      operationId: pauseSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 暂停原因
                  example: "系统维护"
      responses:
        '204':
          description: 暂停成功
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{name}/schedules/{schedule_id}/resume:
    post:
      summary: 恢复调度
      description: |
        恢复已暂停的调度。
        
        恢复后将按 Cron 表达式正常触发。
      tags: [Schedules]
      operationId: resumeSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      responses:
        '204':
          description: 恢复成功
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workflows/{name}/schedules/{schedule_id}/trigger:
    post:
      summary: 手动触发调度
      description: |
        手动触发一次调度执行，不影响正常的 Cron 计划。
        
        可选提供覆盖变量。
      tags: [Schedules]
      operationId: triggerSchedule
      parameters:
        - name: name
          in: path
          required: true
          description: 工作流定义名称
          schema:
            type: string
        - name: schedule_id
          in: path
          required: true
          description: 调度ID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                vars:
                  type: object
                  additionalProperties: true
                  description: 覆盖变量（优先级最高）
                  example:
                    env: "production"
                    timeout: 120
      responses:
        '202':
          description: 触发成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow_id:
                    type: string
                    description: 触发的工作流执行ID
                    example: "wf_20260205_abc123"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ==================== Templates ====================
...
  /v1/templates/{name}:
    get:
      summary: 获取模板内容
      description: 获取指定模板的完整 YAML 内容和元数据
      tags: [Templates]
      operationId: getTemplate
      parameters:
        - name: name
          in: path
          required: true
          description: 模板名称
          schema:
            type: string
      responses:
        '200':
          description: 成功返回模板内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ==================== Nodes ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token 认证
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 认证
...